feat: apy v2

+++ carmine-api-rpc-gateway/src/lib.rs
@@ -186,7 +186,7 @@ struct RpcErrorData {
 #[allow(dead_code)]
 #[derive(Debug, Deserialize)]
 struct RpcErrorResponse {
-    code: u8,
+    code: i64,
     message: String,
     data: RpcErrorData,
 }
